requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
djangorestframework==3.11.1
idna==2.10
passlib==1.7.4
Pillow==8.1.0
pyfcm==1.4.9
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
stripe==2.56.0
urllib3==1.26.2
